Move canmount changes until after reboot.
This means an rc.d script, which has to be in /etc/rc.d not
/usr/local/etc/rc.d since it needs to run before /usr is mounted.

There are several problems with changing canmount before reboot. One is
that if canmount is set to noauto it prevents rc.d/zfs from unmounting
them at shutdown. Another is that more generally zfs u?mount -a will no
longer be trying to mount or unmount the right filesystems.

A third is that booting from an alternative BE is much more difficult
than it needs to be: it's impossible to change the canmount property
from the loader (the loader only has read-only access), so to boot from
an inactive BE you have to boot single-user and change all the canmount
properties by hand. A boot-time script to ensure all the filesystems in
the BE we booted from are active and all other BE filesystems are
inactive is much safer.