Don't copy zpool.cache on activation.
I don't think it's useful. Firstly, zpool.cache is no longer necessary
for booting (if it ever was): zfsloader can find zpools by itself, and
simply ignores the fact that they might not be imported on this machine.
The function of zpool.cache is just to record which pools are imported,
with the root pool always implicitly considered imported.

Secondly, the only way to create a BE with beadm is as a clone of an
existing BE, which means the new BE will inherit a valid zpool.cache (or
none at all, which is valid, and means the only imported pool is the
root pool). Users creating BEs by hand with FreeBSD versions that need
zpool.cache to boot (are there any? I can't find any evidence that there
are...) will have to copy a zpool.cache over themselves, but that's just
a normal part of installing those versions on ZFS. Since they're doing
everything manually, beadm isn't involved yet, and we shoudn't try to
interfere later on.

Thirdly, it seems better, to me, for a 'zpool im/export' *not* to affect
other BEs. That is, booting a BE should make visible the pools that were
imported last time that BE booted (assuming they're available), and not
any pools that happened to be visible in whatever BE we were running
last.

[If/when I add an option to create a new BE from scratch, I need to
consider whether it should be given a zpool.cache.]