Don't touch loader.conf unless we need to.
FreeBSD 10's loader contains code to set vfs.root.mountfrom from currdev
if we are booting from ZFS. This means there's no need to touch
loader.conf on these systems, and in fact it's better if loader.conf
doesn't set vfs.root.mountfrom at all. Check for 10 by looking for
/bin/freebsd-version, and if it's there just issue a warning if
loader.conf seems to be changing vfs.root.mountfrom.

On earlier versions of FreeBSD, update loader.conf when the BE is
created rather than when it's activated. If the user changes it after
that, well, presumably they had a reason for doing so, and it's not
polite to overwrite their changes.

[Strictly speaking, it's the version of the loader that matters, not the
version of the userland, but I think it's reasonable to assume they're
in sync. I still don't like the fact that we're playing with random
files on what may not be a FreeBSD system at all: in an ideal world, it
would be possible to use beadm to switch between FreeBSD and Illumos
BEs, but that won't be possible while there are still FreeBSD versions
that require loader.conf to be manipulated.]